Move CultureValue to OC.Localization module
